Request repair in material and orders

See merge request equipoegresoumss/agronomia-adscripcion-backend!36

+++ app/Http/Controllers/ArticleController.php
@@ -11,11 +11,11 @@ class ArticleController extends Controller
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
     public function index()
     {
-       // $this->authorize('view', User::class);
+        $this->authorize('view', Article::class);
         Article::UpdateStatusIsLow();
         $result = Article::ArticlesAll();
         return $result;
@@ -24,11 +24,12 @@ class ArticleController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
+     * @param  CreateArticleRequest  $request
+     * @return Response
      */
     public function store(CreateArticleRequest $request)
     {
+        $this->authorize('manage', Article::class);
         $article=new Article($request->all());
         $article->stock_total=$article->stock;
         $article->save();
@@ -55,12 +56,13 @@ class ArticleController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param  UpdateArticleRequest  $request
+     * @param  Article  $article
+     * @return Response
      */
     public function update(UpdateArticleRequest $request, Article $article)
     {
+        $this->authorize('manage', Article::class);
         $article->update($request->all());
         return response()->json([
             'res' => true,
@@ -72,10 +74,11 @@ class ArticleController extends Controller
      * Remove the specified resource from storage.
      *
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return Response
      */
     public function destroy($id)
     {
+        $this->authorize('manage', Article::class);
         Article::destroy($id);
         return response()->json([
             'success' => true,
